# openapi.json
{"openapi":"3.1.0","info":{"title":"POPU API","description":"AI-powered epidemiological intelligence backend for Nigeria.","version":"0.1.0"},"paths":{"/api/v1/health":{"get":{"tags":["system"],"summary":"Health","operationId":"health_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Health Api V1 Health Get"}}}}}}},"/api/v1/data/scenario":{"get":{"tags":["epidemiology data"],"summary":"Scenario","operationId":"scenario_api_v1_data_scenario_get","parameters":[{"name":"disease","in":"query","required":true,"schema":{"type":"string","minLength":1,"title":"Disease"}},{"name":"geography","in":"query","required":true,"schema":{"type":"string","minLength":1,"title":"Geography"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScenarioResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/data/scenario-options":{"get":{"tags":["epidemiology data"],"summary":"Scenario Options","operationId":"scenario_options_api_v1_data_scenario_options_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScenarioListResponse"}}}}}}},"/api/v1/analysis/anomaly":{"get":{"tags":["epidemiological analysis"],"summary":"Anomaly","operationId":"anomaly_api_v1_analysis_anomaly_get","parameters":[{"name":"disease","in":"query","required":true,"schema":{"type":"string","minLength":1,"title":"Disease"}},{"name":"geography","in":"query","required":true,"schema":{"type":"string","minLength":1,"title":"Geography"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnomalyResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/analysis/forecast":{"get":{"tags":["epidemiological analysis"],"summary":"Forecast","operationId":"forecast_api_v1_analysis_forecast_get","parameters":[{"name":"disease","in":"query","required":true,"schema":{"type":"string","minLength":1,"title":"Disease"}},{"name":"geography","in":"query","required":true,"schema":{"type":"string","minLength":1,"title":"Geography"}},{"name":"horizon_days","in":"query","required":false,"schema":{"type":"integer","maximum":30,"minimum":1,"default":7,"title":"Horizon Days"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForecastResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/analysis/evidence":{"get":{"tags":["epidemiological analysis"],"summary":"Evidence","operationId":"evidence_api_v1_analysis_evidence_get","parameters":[{"name":"disease","in":"query","required":true,"schema":{"type":"string","minLength":1,"title":"Disease"}},{"name":"geography","in":"query","required":true,"schema":{"type":"string","minLength":1,"title":"Geography"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EvidenceAssessment"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/analysis/risk":{"get":{"tags":["epidemiological analysis"],"summary":"Risk","operationId":"risk_api_v1_analysis_risk_get","parameters":[{"name":"disease","in":"query","required":true,"schema":{"type":"string","minLength":1,"title":"Disease"}},{"name":"geography","in":"query","required":true,"schema":{"type":"string","minLength":1,"title":"Geography"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RiskAssessment"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/analysis/investigation":{"get":{"tags":["epidemiological analysis"],"summary":"Investigation","operationId":"investigation_api_v1_analysis_investigation_get","parameters":[{"name":"disease","in":"query","required":true,"schema":{"type":"string","minLength":1,"title":"Disease"}},{"name":"geography","in":"query","required":true,"schema":{"type":"string","minLength":1,"title":"Geography"}},{"name":"forecast_horizon_days","in":"query","required":false,"schema":{"type":"integer","maximum":30,"minimum":1,"default":7,"title":"Forecast Horizon Days"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvestigationResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/analysis/risk/integrated":{"get":{"tags":["epidemiological analysis"],"summary":"Get Integrated Risk","operationId":"get_integrated_risk_api_v1_analysis_risk_integrated_get","parameters":[{"name":"disease","in":"query","required":true,"schema":{"type":"string","minLength":1,"title":"Disease"}},{"name":"geography","in":"query","required":true,"schema":{"type":"string","minLength":1,"title":"Geography"}},{"name":"forecast_horizon_days","in":"query","required":false,"schema":{"type":"integer","maximum":30,"minimum":1,"default":7,"title":"Forecast Horizon Days"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IntegratedRiskAssessment"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/agent/investigate":{"post":{"tags":["agent"],"summary":"Investigate With Agent","operationId":"investigate_with_agent_api_v1_agent_investigate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/InvestigationResult"},{"$ref":"#/components/schemas/AgentIntent"},{"$ref":"#/components/schemas/DataAvailabilityResponse"}],"title":"Response Investigate With Agent Api V1 Agent Investigate Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/agent/investigate/report":{"post":{"tags":["agent"],"summary":"Generate Agent Investigation Report","operationId":"generate_agent_investigation_report_api_v1_agent_investigate_report_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvestigationReport"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/signals/scan":{"get":{"tags":["signals"],"summary":"Scan Signals","operationId":"scan_signals_api_v1_signals_scan_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/SignalResult"},"type":"array","title":"Response Scan Signals Api V1 Signals Scan Get"}}}}}}},"/api/v1/llm/interpret":{"post":{"tags":["llm"],"summary":"Interpret Investigation","operationId":"interpret_investigation_api_v1_llm_interpret_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LLMInterpretationRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LLMInterpretation"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"tags":["system"],"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Root  Get"}}}}}}}},"components":{"schemas":{"AgentIntent":{"properties":{"intent":{"type":"string","enum":["INVESTIGATE","UNKNOWN"],"title":"Intent"},"disease":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Disease"},"geography":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Geography"},"forecast_horizon_days":{"type":"integer","title":"Forecast Horizon Days","default":7},"original_request":{"type":"string","title":"Original Request"}},"additionalProperties":false,"type":"object","required":["intent","original_request"],"title":"AgentIntent"},"AgentRequest":{"properties":{"request":{"type":"string","minLength":1,"title":"Request"},"forecast_horizon_days":{"type":"integer","maximum":30.0,"minimum":1.0,"title":"Forecast Horizon Days","default":7}},"additionalProperties":false,"type":"object","required":["request"],"title":"AgentRequest"},"AnomalyResult":{"properties":{"status":{"type":"string","enum":["NORMAL","ELEVATED","ANOMALY"],"title":"Status"},"disease":{"type":"string","title":"Disease"},"geography":{"type":"string","title":"Geography"},"data_status":{"type":"string","title":"Data Status"},"notice":{"type":"string","title":"Notice"},"baseline_value":{"type":"number","title":"Baseline Value"},"observed_value":{"type":"number","title":"Observed Value"},"absolute_deviation":{"type":"number","title":"Absolute Deviation"},"percent_deviation":{"type":"number","title":"Percent Deviation"},"threshold_percent":{"type":"number","title":"Threshold Percent"},"method":{"type":"string","title":"Method"},"model_output":{"type":"string","title":"Model Output"},"uncertainty":{"items":{"type":"string"},"type":"array","title":"Uncertainty"},"human_review_required":{"type":"boolean","title":"Human Review Required"}},"additionalProperties":false,"type":"object","required":["status","disease","geography","data_status","notice","baseline_value","observed_value","absolute_deviation","percent_deviation","threshold_percent","method","model_output","uncertainty","human_review_required"],"title":"AnomalyResult"},"DataAvailabilityResponse":{"properties":{"status":{"type":"string","enum":["DATA_UNAVAILABLE","SOURCE_NOT_CONFIGURED","INVALID_GEOGRAPHY"],"title":"Status"},"disease":{"type":"string","title":"Disease"},"geography":{"type":"string","title":"Geography"},"provider":{"type":"string","title":"Provider"},"message":{"type":"string","title":"Message"},"notice":{"type":"string","title":"Notice"}},"additionalProperties":false,"type":"object","required":["status","disease","geography","provider","message","notice"],"title":"DataAvailabilityResponse"},"EvidenceAssessment":{"properties":{"disease":{"type":"string","title":"Disease"},"geography":{"type":"string","title":"Geography"},"data_status":{"type":"string","title":"Data Status"},"notice":{"type":"string","title":"Notice"},"evidence":{"items":{"$ref":"#/components/schemas/EvidenceItem"},"type":"array","title":"Evidence"},"geographic_evidence":{"items":{"$ref":"#/components/schemas/GeographicEvidence"},"type":"array","title":"Geographic Evidence"},"elevated_signal_count":{"type":"integer","title":"Elevated Signal Count"},"moderate_signal_count":{"type":"integer","title":"Moderate Signal Count"},"stable_signal_count":{"type":"integer","title":"Stable Signal Count"},"uncertainty":{"items":{"type":"string"},"type":"array","title":"Uncertainty"},"human_review_required":{"type":"boolean","title":"Human Review Required"}},"additionalProperties":false,"type":"object","required":["disease","geography","data_status","notice","evidence","geographic_evidence","elevated_signal_count","moderate_signal_count","stable_signal_count","uncertainty","human_review_required"],"title":"EvidenceAssessment"},"EvidenceItem":{"properties":{"source":{"type":"string","title":"Source"},"signal":{"type":"string","enum":["STABLE","MODERATE","ELEVATED","UNKNOWN"],"title":"Signal"},"summary":{"type":"string","title":"Summary"},"observed_value":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Observed Value"},"data_status":{"type":"string","title":"Data Status"},"interpretation_type":{"type":"string","title":"Interpretation Type"},"uncertainty":{"items":{"type":"string"},"type":"array","title":"Uncertainty"}},"additionalProperties":false,"type":"object","required":["source","signal","summary","data_status","interpretation_type","uncertainty"],"title":"EvidenceItem"},"ForecastResult":{"properties":{"disease":{"type":"string","title":"Disease"},"geography":{"type":"string","title":"Geography"},"forecast_horizon_days":{"type":"integer","title":"Forecast Horizon Days"},"predicted_values":{"items":{"type":"number"},"type":"array","title":"Predicted Values"},"risk_level":{"type":"string","enum":["LOW","MODERATE","ELEVATED"],"title":"Risk Level"},"risk_score":{"type":"number","title":"Risk Score"},"confidence_interval":{"items":{"additionalProperties":{"type":"number"},"type":"object"},"type":"array","title":"Confidence Interval"},"model":{"type":"string","title":"Model"},"model_version":{"type":"string","title":"Model Version"},"data_status":{"type":"string","title":"Data Status"},"notice":{"type":"string","title":"Notice"},"uncertainty":{"items":{"type":"string"},"type":"array","title":"Uncertainty"},"human_review_required":{"type":"boolean","title":"Human Review Required"}},"additionalProperties":false,"type":"object","required":["disease","geography","forecast_horizon_days","predicted_values","risk_level","risk_score","confidence_interval","model","model_version","data_status","notice","uncertainty","human_review_required"],"title":"ForecastResult"},"GeographicEvidence":{"properties":{"geography":{"type":"string","title":"Geography"},"cases":{"type":"number","title":"Cases"},"baseline_cases":{"type":"number","title":"Baseline Cases"},"signal":{"type":"string","enum":["STABLE","MODERATE","ELEVATED","UNKNOWN"],"title":"Signal"},"deviation_percent":{"type":"number","title":"Deviation Percent"}},"additionalProperties":false,"type":"object","required":["geography","cases","baseline_cases","signal","deviation_percent"],"title":"GeographicEvidence"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IntegratedRiskAssessment":{"properties":{"disease":{"type":"string","title":"Disease"},"geography":{"type":"string","title":"Geography"},"risk_level":{"type":"string","enum":["LOW","MODERATE","ELEVATED"],"title":"Risk Level"},"risk_score":{"type":"number","title":"Risk Score"},"components":{"items":{"$ref":"#/components/schemas/RiskComponent"},"type":"array","title":"Components"},"key_factors":{"items":{"type":"string"},"type":"array","title":"Key Factors"},"data_status":{"type":"string","title":"Data Status"},"notice":{"type":"string","title":"Notice"},"uncertainty":{"items":{"type":"string"},"type":"array","title":"Uncertainty"},"human_review_required":{"type":"boolean","title":"Human Review Required"}},"additionalProperties":false,"type":"object","required":["disease","geography","risk_level","risk_score","components","key_factors","data_status","notice","uncertainty","human_review_required"],"title":"IntegratedRiskAssessment"},"InvestigationReport":{"properties":{"disease":{"type":"string","title":"Disease"},"geography":{"type":"string","title":"Geography"},"summary":{"type":"string","title":"Summary"},"key_findings":{"items":{"type":"string"},"type":"array","title":"Key Findings"},"investigation_priorities":{"items":{"type":"string"},"type":"array","title":"Investigation Priorities"},"supporting_evidence":{"items":{"type":"string"},"type":"array","title":"Supporting Evidence"},"uncertainty":{"items":{"type":"string"},"type":"array","title":"Uncertainty"},"data_status":{"type":"string","title":"Data Status"},"notice":{"type":"string","title":"Notice"},"human_review_required":{"type":"boolean","title":"Human Review Required"}},"additionalProperties":false,"type":"object","required":["disease","geography","summary","key_findings","investigation_priorities","supporting_evidence","uncertainty","data_status","notice","human_review_required"],"title":"InvestigationReport"},"InvestigationResult":{"properties":{"disease":{"type":"string","title":"Disease"},"geography":{"type":"string","title":"Geography"},"anomaly":{"$ref":"#/components/schemas/AnomalyResult"},"evidence":{"$ref":"#/components/schemas/EvidenceAssessment"},"forecast":{"$ref":"#/components/schemas/ForecastResult"},"risk":{"$ref":"#/components/schemas/RiskAssessment"},"integrated_risk":{"$ref":"#/components/schemas/IntegratedRiskAssessment"},"overall_signal":{"type":"string","title":"Overall Signal"},"key_findings":{"items":{"type":"string"},"type":"array","title":"Key Findings"},"data_status":{"type":"string","title":"Data Status"},"notice":{"type":"string","title":"Notice"},"uncertainty":{"items":{"type":"string"},"type":"array","title":"Uncertainty"},"human_review_required":{"type":"boolean","title":"Human Review Required"}},"additionalProperties":false,"type":"object","required":["disease","geography","anomaly","evidence","forecast","risk","integrated_risk","overall_signal","key_findings","data_status","notice","uncertainty","human_review_required"],"title":"InvestigationResult"},"LLMInterpretation":{"properties":{"provider":{"type":"string","title":"Provider"},"model":{"type":"string","title":"Model"},"fallback_used":{"type":"boolean","title":"Fallback Used"},"interpretation":{"type":"string","title":"Interpretation"},"uncertainty":{"items":{"type":"string"},"type":"array","title":"Uncertainty"},"human_review_required":{"type":"boolean","title":"Human Review Required"}},"additionalProperties":false,"type":"object","required":["provider","model","fallback_used","interpretation","uncertainty","human_review_required"],"title":"LLMInterpretation"},"LLMInterpretationRequest":{"properties":{"request":{"type":"string","minLength":3,"title":"Request"},"forecast_horizon_days":{"type":"integer","maximum":30.0,"minimum":1.0,"title":"Forecast Horizon Days","default":14}},"additionalProperties":false,"type":"object","required":["request"],"title":"LLMInterpretationRequest"},"RiskAssessment":{"properties":{"disease":{"type":"string","title":"Disease"},"geography":{"type":"string","title":"Geography"},"risk_level":{"type":"string","enum":["LOW","MODERATE","ELEVATED"],"title":"Risk Level"},"risk_score":{"type":"number","title":"Risk Score"},"evidence_signal_count":{"type":"integer","title":"Evidence Signal Count"},"elevated_signal_count":{"type":"integer","title":"Elevated Signal Count"},"moderate_signal_count":{"type":"integer","title":"Moderate Signal Count"},"stable_signal_count":{"type":"integer","title":"Stable Signal Count"},"contributing_factors":{"items":{"type":"string"},"type":"array","title":"Contributing Factors"},"data_status":{"type":"string","title":"Data Status"},"notice":{"type":"string","title":"Notice"},"uncertainty":{"items":{"type":"string"},"type":"array","title":"Uncertainty"},"human_review_required":{"type":"boolean","title":"Human Review Required"}},"additionalProperties":false,"type":"object","required":["disease","geography","risk_level","risk_score","evidence_signal_count","elevated_signal_count","moderate_signal_count","stable_signal_count","contributing_factors","data_status","notice","uncertainty","human_review_required"],"title":"RiskAssessment"},"RiskComponent":{"properties":{"name":{"type":"string","title":"Name"},"signal":{"type":"string","title":"Signal"},"contribution":{"type":"number","title":"Contribution"},"explanation":{"type":"string","title":"Explanation"}},"additionalProperties":false,"type":"object","required":["name","signal","contribution","explanation"],"title":"RiskComponent"},"ScenarioListResponse":{"properties":{"diseases":{"items":{"type":"string"},"type":"array","title":"Diseases"},"geographies":{"items":{"type":"string"},"type":"array","title":"Geographies"},"data_status":{"type":"string","title":"Data Status","default":"SYNTHETIC DATA"},"notice":{"type":"string","title":"Notice"}},"type":"object","required":["diseases","geographies","notice"],"title":"ScenarioListResponse"},"ScenarioResponse":{"properties":{"status":{"type":"string","title":"Status"},"provider":{"type":"string","title":"Provider"},"mode":{"type":"string","title":"Mode"},"disease":{"type":"string","title":"Disease"},"geography":{"type":"string","title":"Geography"},"data_status":{"type":"string","title":"Data Status"},"notice":{"type":"string","title":"Notice"},"scenario":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Scenario"},"message":{"type":"string","title":"Message"}},"additionalProperties":false,"type":"object","required":["status","provider","mode","disease","geography","data_status","notice","message"],"title":"ScenarioResponse"},"SignalResult":{"properties":{"signal_id":{"type":"string","title":"Signal Id"},"disease":{"type":"string","title":"Disease"},"geography":{"type":"string","title":"Geography"},"status":{"type":"string","enum":["ANOMALY","ELEVATED"],"title":"Status"},"severity":{"type":"string","enum":["HIGH","MODERATE"],"title":"Severity"},"detected_at":{"type":"string","format":"date-time","title":"Detected At"},"baseline_value":{"type":"number","title":"Baseline Value"},"observed_value":{"type":"number","title":"Observed Value"},"percent_deviation":{"type":"number","title":"Percent Deviation"},"detection_method":{"type":"string","title":"Detection Method"},"source":{"type":"string","title":"Source"},"summary":{"type":"string","title":"Summary"},"supporting_sources":{"items":{"type":"string"},"type":"array","title":"Supporting Sources"},"elevated_source_count":{"type":"integer","title":"Elevated Source Count","default":0},"moderate_source_count":{"type":"integer","title":"Moderate Source Count","default":0},"geographic_signal_count":{"type":"integer","title":"Geographic Signal Count","default":0},"risk_level":{"type":"string","enum":["LOW","MODERATE","ELEVATED"],"title":"Risk Level"},"risk_score":{"type":"number","title":"Risk Score"},"forecast_risk_level":{"type":"string","enum":["LOW","MODERATE","ELEVATED"],"title":"Forecast Risk Level"},"forecast_risk_score":{"type":"number","title":"Forecast Risk Score"},"forecast_horizon_days":{"type":"integer","title":"Forecast Horizon Days"},"forecast_direction":{"type":"string","enum":["INCREASING","STABLE","DECREASING"],"title":"Forecast Direction"},"data_status":{"type":"string","title":"Data Status"},"notice":{"type":"string","title":"Notice"},"human_review_required":{"type":"boolean","title":"Human Review Required","default":true},"uncertainty":{"items":{"type":"string"},"type":"array","title":"Uncertainty"}},"type":"object","required":["signal_id","disease","geography","status","severity","detected_at","baseline_value","observed_value","percent_deviation","detection_method","source","summary","risk_level","risk_score","forecast_risk_level","forecast_risk_score","forecast_horizon_days","forecast_direction","data_status","notice"],"title":"SignalResult"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}